+++ libs/clients/bff/src/api.ts
@@ -17,6 +17,7 @@
 export * from './api/default-api';
 export * from './api/remote-catalogue-api';
 export * from './api/remote-group-api';
+export * from './api/remote-kitchen-api';
 export * from './api/remote-offer-api';
 export * from './api/remote-table-api';