+++ libs/services/common/src/lib/kitchen/kitchenService.ts
@@ -1,11 +1,27 @@
-import { OrderSummary } from './kitchenServiceWorkflow';
+
+export interface CartItem {
+  itemId: string;
+  shortName: string;
+  quantity: number;
+}
 
 export interface MonsieurAxelMenvoie {
-  cart: { itemId: string, shortName: string; quantity: number }[];
+  cart: CartItem[];
   groupId: string;
   tableNumber: number;
 }
+export interface PreparationStatus {
+  status:  "readyToBeServed" | "preparationStarted" | "preparationServed";
+  preparationId: string;
+}
 
+export interface OrderSummary {
+  summary: {
+    [category: string]: {
+      [table: number]: PreparationStatus[];
+    };
+  };
+}
 export interface KitchenService {
   sendToKitchen(order: MonsieurAxelMenvoie): Promise<void>;
   getOrdersByGroupId(groupId: string): Promise<OrderSummary>;